Created module for preprocessing growth_matrices Created module for handling ticket splits Created module for general utils (mainly matrix manipulations and conversions) Deleted functions from edge_replicant as they're moved to new modules Nothing new tested yet
