+++ normits_demand/models/forecasting/edge_replicant.py
@@ -28,278 +28,6 @@ LOG = logging.getLogger(__name__)
 # ## FUNCTIONS ## #
 
 
-def long_mx_2_wide_mx(
-    mx_df: pd.DataFrame,
-    row: str = "no_entry",
-    col: str = "no_entry",
-    value: str = "no_entry",
-) -> np.ndarray:
-    """Convert pandas long matrix to numpy wide matrix.
-
-    Function assumes default entry of a pandas dataframe of three columns:
-        [origin/from/production, destination,to/attraction, demand]
-    Can be specified if the matrix is of a different length or order.
-
-    Parameters
-    ----------
-    mx_df : pd.DataFrame
-        pandas long matrix dataframe to convert
-    row : str, optional
-        rows vector in the matrix dataframe
-    col : str, optional
-        columns vector in the matrix dataframe
-    value : str, optional
-        demand vector in the matrix dataframe
-
-    Returns
-    -------
-    np.ndarray
-        numpy wide matrix
-    """
-    # if user specified entries
-    if row == "no_entry":
-        row = mx_df.columns[0]
-    if col == "no_entry":
-        col = mx_df.columns[1]
-    if value == "no_entry":
-        value = mx_df.columns[2]
-
-    # reshape to wide numpy matrix
-    wide_mx = mx_df.pivot_table(index=row, columns=col, values=value).values
-
-    return wide_mx
-
-
-def wide_mx_2_long_mx(
-    mx_np: np.ndarray,
-    rows: str = "from_stn_zone_id",
-    cols: str = "to_stn_zone_id",
-    values: str = "Demand",
-) -> pd.DataFrame:
-    """Convert numpy wide matrix to pandas long matrix.
-
-    Function assumes conversion is happening to a stn2stn matrix hence the headers
-    for the output dataframe are named to station level by default. Optional entries
-    can be given through rows, cols and values
-
-    Parameters
-    ----------
-    mx_np : np.ndarray
-        numpy wide matrix dataframe to convert
-    row : str, optional
-        rows vector in the matrix dataframe
-    col : str, optional
-        columns vector in the matrix dataframe
-    value : str, optional
-        demand vector in the matrix dataframe
-
-    Returns
-    -------
-    mx_df : pd.DataFrame
-        pandas long matrix
-    """
-    # get omx array to pandas dataframe and reset productions
-    mx_df = pd.DataFrame(mx_np).reset_index().rename(columns={"index": rows})
-    # melt DF to get attractions vector
-    mx_df = mx_df.melt(id_vars=[rows], var_name=cols, value_name=values)
-    # adjust zone number
-    mx_df[rows] = mx_df[rows] + 1
-    mx_df[cols] = mx_df[cols] + 1
-
-    return mx_df
-
-
-def transpose_matrix(mx_df: pd.DataFrame, stations: bool = False) -> pd.DataFrame:
-    """Transpose a matrix O<>D/P<>A.
-
-    Parameters
-    ----------
-    mx : pd.DataFrame
-        input matrix to transpose
-    stations : bool
-        whether it's a stations matrix or not
-
-    Returns
-    -------
-    mx : pd.DataFrame
-        transposed matrix
-    """
-    # o/d columns
-    from_col = "from_model_zone_id"
-    to_col = "to_model_zone_id"
-    # if stations matrix, update the od columns
-    if stations:
-        from_col = "from_stn_zone_id"
-        to_col = "to_stn_zone_id"
-    # transpose to-home PA to OD by renaming from <> to model zone id
-    mx_df = mx_df.rename(
-        columns={
-            from_col: to_col,
-            to_col: from_col,
-        }
-    )
-
-    return mx_df
-
-
-def expand_matrix(
-    mx_df: pd.DataFrame, zones: int = 1300, stations: bool = False
-) -> pd.DataFrame:
-    """Expand matrix to all possible movements (zones x zones).
-
-    Parameters
-    ----------
-    mx_df : pd.DataFrame
-        matrix
-    zones: int
-        number of model zones, default = 1300
-    stations : bool
-        whether it's a stations matrix or not'
-    Returns
-    -------
-    expanded_mx : pd.DataFrame
-        expanded matrix
-    """
-    # o/d columns
-    od_cols = ["from_model_zone_id", "to_model_zone_id"]
-    # if stations matrix, update the od columns
-    if stations:
-        od_cols = ["from_stn_zone_id", "to_stn_zone_id"]
-    # create empty dataframe
-    expanded_mx = pd.DataFrame(
-        list(itertools.product(range(1, zones + 1), range(1, zones + 1))),
-        columns=od_cols,
-    )
-    # get first matrix
-    expanded_mx = expanded_mx.merge(mx_df, how="outer", on=od_cols).fillna(0)
-    return expanded_mx
-
-
-def add_distance_band_tag_flow(mx_df: pd.DataFrame) -> pd.DataFrame:
-    """Add TAGs distance band based on distance.
-
-    Parameters
-    ----------
-    mx_df : pd.DataFrame
-        prepared matrix with flows
-
-    Returns
-    -------
-    mx_df : pd.DataFrame
-        dataframe with added new TAG flow
-    """
-    # set new flow to match the non-distance flow to begin with
-    mx_df.loc[:, "TAG_Flow"] = mx_df["TAG_NonDist"]
-    # Outside South East
-    mx_df.loc[
-        (mx_df["TAG_NonDist"] == "Outside South East".lower()) & (mx_df["Distance"] < 25),
-        "TAG_Flow",
-    ] = "Outside South East <25 miles".lower()
-    mx_df.loc[
-        (mx_df["TAG_NonDist"] == "Outside South East".lower())
-        & (mx_df["Distance"] >= 25)
-        & (mx_df["Distance"] < 100),
-        "TAG_Flow",
-    ] = "Outside South East 25 to 100 miles".lower()
-    mx_df.loc[
-        (mx_df["TAG_NonDist"] == "Outside South East".lower()) & (mx_df["Distance"] >= 100),
-        "TAG_Flow",
-    ] = "Outside South East  100 + miles - adjusted".lower()
-    # Outside South East to/from London
-    mx_df.loc[
-        (mx_df["TAG_NonDist"] == "Outside South East to/from London".lower())
-        & (mx_df["Distance"] < 100),
-        "TAG_Flow",
-    ] = "Outside South East to/from London < 100 miles".lower()
-    mx_df.loc[
-        (mx_df["TAG_NonDist"] == "Outside South East to/from London".lower())
-        & (mx_df["Distance"] >= 100),
-        "TAG_Flow",
-    ] = "Outside South East to/from London 100 + miles".lower()
-
-    return mx_df
-
-
-def prepare_growth_matrices(
-    demand_segments_df: pd.DataFrame,
-    factors_df: pd.DataFrame,
-    stations_lookup: pd.DataFrame,
-) -> dict:
-    """Prepare Growth Factor matrices into numpy matrices.
-
-    Function creates growth numpy matrix for all purposes and ticket types combination
-    on a station to station level.
-
-    Parameters
-    ----------
-    demand_segments_df : pd.DataFrame
-        models' demand segments information dataframe
-    factors_df : pd.DataFrame
-        EDGE growth factors dataframe
-    stations_lookup : pd.DataFrame
-        lookup for all model used stations zones and TLCs
-
-    Returns
-    -------
-    growth_matrices : dict
-        numpy growth matrices for all purposes and ticket types
-    """
-    # get list of purposes
-    purposes = demand_segments_df["Purpose"].drop_duplicates().to_list()
-    # get list of ticket types
-    ticket_types = factors_df["TicketType"].drop_duplicates().to_list()
-    # create a list of model used stations
-    used_stations = stations_lookup["STATIONCODE"].to_list()
-    # filter factors file to keep only used stations
-    factors_df = factors_df.loc[
-        (factors_df["ZoneCodeFrom"].isin(used_stations))
-        & (factors_df["ZoneCodeTo"].isin(used_stations))
-    ]
-
-    # add stns zones
-    # merge on origin/production
-    factors_df = factors_df.merge(
-        stations_lookup, how="left", left_on=["ZoneCodeFrom"], right_on=["STATIONCODE"]
-    )
-    # rename
-    factors_df = factors_df.rename(columns={"stn_zone_id": "from_stn_zone_id"})
-    # merge on destination/attraction
-    factors_df = factors_df.merge(
-        stations_lookup, how="left", left_on=["ZoneCodeTo"], right_on=["STATIONCODE"]
-    )
-    # rename
-    factors_df = factors_df.rename(
-        columns={"stn_zone_id": "to_stn_zone_id", "Demand_rate": "Demand"}
-    )
-    # keep needed columns
-    factors_df = factors_df[
-        [
-            "from_stn_zone_id",
-            "to_stn_zone_id",
-            "ZoneCodeFrom",
-            "ZoneCodeTo",
-            "purpose",
-            "TicketType",
-            "Demand",
-        ]
-    ]
-
-    # create matrices dictionary
-    growth_matrices = {}
-
-    # get growth matrices for each purpose/ticket type
-    for purpose in purposes:
-        for ticket_type in ticket_types:
-            mx_df = factors_df[["from_stn_zone_id", "to_stn_zone_id", "Demand"]].loc[
-                (factors_df["purpose"] == purpose) & (factors_df["TicketType"] == ticket_type)
-            ]
-            # expand matrix
-            mx_df = expand_matrix(mx_df, zones=len(stations_lookup), stations=True)
-            growth_matrices[f"{purpose}_{ticket_type}"] = long_mx_2_wide_mx(mx_df)
-
-    return growth_matrices
-
-
 def zonal_from_to_stations_demand(
     demand_mx: pd.DataFrame,
     irsj_props: pd.DataFrame,
@@ -388,165 +116,6 @@ def zonal_from_to_stations_demand(
     np_mx = long_mx_2_wide_mx(mx_df)
 
     return np_mx, zonal_from_to_stns
-
-
-def produce_ticketype_splitting_matrices(
-    edge_flows: pd.DataFrame,
-    stations_lookup: pd.DataFrame,
-    distance_mx: pd.DataFrame,
-    flows_lookup: pd.DataFrame,
-    ticket_split_proportions: pd.DataFrame,
-) -> dict:
-    """Produce numpy matrices to split stn2stn demand into the different ticket types.
-
-    Parameters
-    ----------
-    edge_flows : pd.DataFrame
-        all eDGE flows per station pair dataframe
-    stations_lookup : pd.DataFrame
-        lookup for all model used stations zones and TLCs
-    distance_mx : pd.DataFrame
-        stn2stn distance matrix
-    flows_lookup : pd.DataFrame
-        lookup for flow category names between EDGE and TAG
-    ticket_split_proportions : pd.DataFrame
-        ticket type splitting proportions dataframe
-
-    Returns
-    -------
-    splitting_matrices : dict
-        numpy ticket type splitting matrices for all purposes and ticket types
-    """
-    # normalize flows
-    edge_flows.loc[:, "FlowCatName"] = edge_flows["FlowCatName"].str.lower()
-    flows_lookup.loc[:, "FlowCatName"] = flows_lookup["FlowCatName"].str.lower()
-    flows_lookup.loc[:, "TAG_NonDist"] = flows_lookup["TAG_NonDist"].str.lower()
-    ticket_split_proportions.loc[:, "TAG_Flow"] = ticket_split_proportions[
-        "TAG_Flow"
-    ].str.lower()
-
-    # create a list of model used stations
-    used_stations = stations_lookup["STATIONCODE"].to_list()
-    # filter flows to keep only used stations
-    edge_flows = edge_flows.loc[
-        (edge_flows["FromCaseZoneID"].isin(used_stations))
-        & (edge_flows["ToCaseZoneID"].isin(used_stations))
-    ]
-    # add flows TAG category
-    edge_flows = edge_flows.merge(flows_lookup, how="left", on=["FlowCatName"])
-    # add station zones
-    # origin/production
-    edge_flows = edge_flows.merge(
-        stations_lookup, how="left", left_on=["FromCaseZoneID"], right_on=["STATIONCODE"]
-    )
-    edge_flows = edge_flows.rename(columns={"stn_zone_id": "from_stn_zone_id"})
-    # destination/attraction
-    edge_flows = edge_flows.merge(
-        stations_lookup, how="left", left_on=["ToCaseZoneID"], right_on=["STATIONCODE"]
-    )
-    edge_flows = edge_flows.rename(columns={"stn_zone_id": "to_stn_zone_id"})
-    # keep needed columns
-    edge_flows = edge_flows[
-        [
-            "from_stn_zone_id",
-            "to_stn_zone_id",
-            "FromCaseZoneID",
-            "ToCaseZoneID",
-            "FlowCatName",
-            "TAG_NonDist",
-        ]
-    ]
-    # merge distance to flows
-    edge_flows = edge_flows.merge(
-        distance_mx, how="left", on=["from_stn_zone_id", "to_stn_zone_id"]
-    )
-    # rename
-    edge_flows = edge_flows.rename(columns={"tran_distance": "Distance"})
-    # fill na
-    edge_flows = edge_flows.fillna(0)
-    # allocate distance bands
-    edge_flows = add_distance_band_tag_flow(edge_flows)
-    # keep needed columns
-    edge_flows = edge_flows[
-        ["from_stn_zone_id", "to_stn_zone_id", "FromCaseZoneID", "ToCaseZoneID", "TAG_Flow"]
-    ]
-    # merge ticket split factors
-    edge_flows = edge_flows.merge(ticket_split_proportions, how="left", on=["TAG_Flow"])
-    # get list of purposes
-    purposes = edge_flows["Purpose"].drop_duplicates().to_list()
-    # create matrices dictionary
-    splitting_matrices = {}
-    # create numpy splitting matrices
-    for purpose in purposes:
-        for ticketype in ["F", "R", "S"]:
-            # get current purpose
-            mx_df = edge_flows.loc[edge_flows["Purpose"] == purpose].reset_index(drop=True)
-            # keep needed columns
-            mx_df = mx_df[["from_stn_zone_id", "to_stn_zone_id", ticketype]]
-            # rename
-            mx_df = mx_df.rename(columns={ticketype: "Demand"})
-            # expand matrix
-            mx_df = expand_matrix(mx_df, zones=len(stations_lookup), stations=True)
-            # convert to numpy and add to matrices dictionary
-            splitting_matrices[f"{purpose}_{ticketype}"] = long_mx_2_wide_mx(mx_df)
-
-    return splitting_matrices
-
-
-def fill_missing_factors(
-    purposes: list,
-    growth_matrices: dict,
-) -> dict:
-    """Fills missing factors for specific ticket type with an available one.
-
-    The filling process looks for an available factor for the same station pair
-    and journey purpose and follows the below hierarchy in filling missing factors:
-
-        Full tickets: First look for Reduced and then look for Season
-        Reduced tickets: First look for Full and then look for Season
-        Season tickets: First look for Reduced and then look for Full
-
-    function also adds a growth factor of 1 (i.e. no growth) where no factor is available
-
-    Parameters
-    ----------
-    purposes : list
-        list of journey purposes
-    growth_matrices : dict
-        numpy growth matrices for all journey purposes and ticket types
-
-    Returns
-    -------
-    filled_growth_matrices : dict
-        filled growth factor matrices
-    """
-    # order S: R, F
-    filled_growth_matrices = {}
-    for purpose in purposes:
-        # get current matrices
-        f_mx = growth_matrices[f"{purpose}_F"]
-        r_mx = growth_matrices[f"{purpose}_R"]
-        s_mx = growth_matrices[f"{purpose}_S"]
-        # create a new growth factors matrix and fill from other ticket types
-        # full - order F > R > S
-        filled_f_mx = np.where(f_mx == 0, r_mx, f_mx)
-        filled_f_mx = np.where(filled_f_mx == 0, s_mx, filled_f_mx)
-        filled_f_mx = np.where(filled_f_mx == 0, 1, filled_f_mx)
-        # reduced - order R > F > S
-        filled_r_mx = np.where(r_mx == 0, f_mx, r_mx)
-        filled_r_mx = np.where(filled_r_mx == 0, s_mx, filled_r_mx)
-        filled_r_mx = np.where(filled_r_mx == 0, 1, filled_r_mx)
-        # season - order S > R > F
-        filled_s_mx = np.where(s_mx == 0, r_mx, s_mx)
-        filled_s_mx = np.where(filled_s_mx == 0, f_mx, filled_s_mx)
-        filled_s_mx = np.where(filled_s_mx == 0, 1, filled_s_mx)
-
-        # append to filled matrices
-        filled_growth_matrices[f"{purpose}_F"] = filled_f_mx
-        filled_growth_matrices[f"{purpose}_R"] = filled_r_mx
-        filled_growth_matrices[f"{purpose}_S"] = filled_s_mx
-
-    return filled_growth_matrices
 
 
 def apply_demand_growth(
@@ -711,38 +280,7 @@ def fromto_2_from_by_averaging(
     return matrices
 
 
-def convert_csv_2_mat(
-    norms_segments: list,
-    cube_exe: pathlib.Path,
-    fcast_year: int,
-    output_folder: pathlib.Path,
-) -> None:
-    """Convert CSV output matrices to Cube .MAT.
-    Function converts output CSV matrices into a single Cube .MAT matrix
-    in NoRMS input demand matrix format
-    Parameters
-    ----------
-    norms_segments : list
-        list of NoRMS input demand segments
-    cube_exe : Path
-        path to Cube Voyager executable
-    fcast_year : int
-        forecast year
-    output_folder : Path
-        path to folder where CSV matrices are saved. this is where the .MAT
-        will also be saved to
-    """
-    # empty dictionary
-    mats_dict = {}
-    # create a dictionary of matrices and their paths
-    for segment in norms_segments:
-        mats_dict[segment] = pathlib.Path(output_folder, f"{fcast_year}_24Hr_{segment}.csv")
 
-    # call CUBE convertor class
-    c_m = CUBEMatConverter(cube_exe)
-    c_m.csv_to_mat(
-        1300, mats_dict, pathlib.Path(output_folder, f"PT_24hr_Demand_{fcast_year}.MAT"), 1
-    )
 
 
 # TODO (MI): Possibly modularize this into 2 or 3 modules. e.g. growth matrices, splitting matrices, growth application
